# perf: combine visitors