--- a/src/combineVisitors.js
+++ b/src/combineVisitors.js
@@ -4,87 +4,83 @@ import * as t from '@babel/types';
 /**
  * Combines consecutive visitor objects to improve performance by reducing overhead.
  * @warning `CallExpression` is filtered and only called for methods listed in `functionNames`.
- * @param rawVisitors An array of functions and/or visitor objects
+ * @param visitors An array of functions and/or visitor objects
  */
-export default function combineVisitors(rawVisitors) {
-  const processedVisitors = [];
-  let tempVisitorHolder = [];
-
-  // Combine consecutive visitor objects into an array
-  for (const visitor of rawVisitors) {
-    if (typeof visitor === 'object') {
-      tempVisitorHolder.push(visitor);
-      continue;
-    }
-
-    if (tempVisitorHolder.length) {
-      processedVisitors.push(tempVisitorHolder);
-      tempVisitorHolder = [];
-    }
-    processedVisitors.push(visitor);
-  }
-
-  if (tempVisitorHolder.length) {
-    // The last item in visitors was an object
-    processedVisitors.push(tempVisitorHolder);
-  }
-
-  return processedVisitors.map(visitor => {
-    if (!Array.isArray(visitor)) {
-      return visitor;
-    }
-
-    /**
-     * { CallExpression: [visitor1, visitor2, ...], ...}
-     */
-    const visitorLookup = {};
-    /**
-     * {
-     * CallExpression(...args) {
-     *  visitorLookup[CallExpression].forEach(x => x.call(this, ...args))
-     * }
-     */
-    let mainVisitor = {};
-
-    for (const obj of visitor) {
-      _.forOwn(obj, (value, key) => {
-        // Add the visitor to the lookup so it can be called from the mainVisitor
-        if (visitorLookup[key] === undefined) {
-          visitorLookup[key] = [];
+export default function combineVisitors(visitors) {
+  return (
+    visitors
+      // Combine consecutive visitor objects into an array
+      .reduce((accumulator, visitor) => {
+        if (typeof visitor === 'object') {
+          if (accumulator.length && Array.isArray(accumulator[accumulator.length - 1])) {
+            accumulator[accumulator.length - 1].push(visitor);
+          } else {
+            accumulator.push([visitor]);
+          }
+        } else {
+          accumulator.push(visitor);
         }
-        visitorLookup[key].push(value);
-
-        if (mainVisitor[key] !== undefined) {
-          return;
+        return accumulator;
+      }, [])
+      // Convert arrays of consecutive visitor objects to a single visitor object that calls
+      // all visitors in the array
+      .map(visitor => {
+        if (!Array.isArray(visitor)) {
+          return visitor;
         }
 
-        // Create the visitor that will be called by babel
-        mainVisitor = {
-          ...mainVisitor,
-          [key](path, ...args) {
-            // Skip CallExpression visitors if it isn't clsx, classnames, etc.
-            if (key === 'CallExpression') {
-              const c = path.node.callee;
-              if (!t.isIdentifier(c) || !this.options.functionNames.includes(c.name)) {
-                return;
-              }
+        /**
+         * { CallExpression: [visitor1, visitor2, ...], ...}
+         */
+        const visitorLookup = {};
+        /**
+         * {
+         * CallExpression(...args) {
+         *  visitorLookup[CallExpression].forEach(x => x.call(this, ...args))
+         * }
+         */
+        let combinedVisitors = {};
+
+        for (const obj of visitor) {
+          _.forOwn(obj, (value, key) => {
+            // Add the visitor to the lookup so it can be called from the mainVisitor
+            if (visitorLookup[key] === undefined) {
+              visitorLookup[key] = [];
+            }
+            visitorLookup[key].push(value);
+
+            if (combinedVisitors[key] !== undefined) {
+              return;
             }
 
-            for (const x of visitorLookup[key]) {
-              x.call(this, path, ...args);
-              // If the visitor calls path.skip, path.remove, or the type changes (path.replace*)
-              // then skip all other visitors in the group
-              if (path.shouldSkip || path.type !== key) {
-                return;
-              }
-            }
-          },
+            // Create the visitor that will be called by babel
+            combinedVisitors = {
+              ...combinedVisitors,
+              [key](path, ...args) {
+                // Skip CallExpression visitors if it isn't clsx, classnames, etc.
+                if (key === 'CallExpression') {
+                  const c = path.node.callee;
+                  if (!t.isIdentifier(c) || !this.options.functionNames.includes(c.name)) {
+                    return;
+                  }
+                }
+
+                for (const x of visitorLookup[key]) {
+                  x.call(this, path, ...args);
+                  // If the visitor calls path.skip, path.remove, or the type changes (path.replace*)
+                  // then skip all other visitors in the group
+                  if (path.shouldSkip || path.type !== key) {
+                    return;
+                  }
+                }
+              },
+            };
+          });
+        }
+
+        return (path, options) => {
+          path.traverse(combinedVisitors, { options });
         };
-      });
-    }
-
-    return (path, options) => {
-      path.traverse(mainVisitor, { options });
-    };
-  });
+      })
+  );
 }