Combine visitors to improve performance by reducing overhead

Time spent optimizing clsx calls

||before|after|change|
|--|--|--|--|
|[material-ui docs](https://github.com/mui-org/material-ui/tree/master/docs) | 1526ms |890ms|-636ms (-41.68%)|
|[@material-ui/core](https://github.com/mui-org/material-ui/tree/master/packages/material-ui)|796ms|515ms|-281ms (-35.3%)|